refactor: make team service less visible
